Move FSNodeEntry and FSNodeType into filescanner.py
